add standalone test for web uploading - https://github.com/emsesp/EMS-ESP32/issues/1564
