+++ interface/package.json
@@ -14,8 +14,9 @@
     "preview-standalone": "typesafe-i18n --no-watch && vite build && concurrently -c \"auto\" \"npm:mock-api\" \"vite preview\"",
     "mock-api": "bun --watch ../mock-api/rest_server.ts",
     "mock-es": "bun --watch ../mock-api/es_server.ts",
+    "mock-upload": "bun --watch ../mock-api/upload_server.ts",
     "old_mock-api": "bun --watch ../mock-api/server.js",
-    "standalone": "concurrently -c \"auto\" \"typesafe-i18n\" \"npm:mock-api\" \"npm:mock-es\" \"vite\"",
+    "standalone": "concurrently -c \"auto\" \"typesafe-i18n\" \"npm:mock-api\" \"npm:mock-es\" \"npm:mock-upload\" \"vite\"",
     "old_standalone": "concurrently -c \"auto\" \"typesafe-i18n\" \"npm:old_mock-api\" \"vite\"",
     "typesafe-i18n": "typesafe-i18n --no-watch",
     "webUI": "node progmem-generator.js",
@@ -24,19 +25,19 @@
   },
   "dependencies": {
     "@alova/adapter-xhr": "^1.0.3",
-    "@babel/core": "^7.24.3",
+    "@babel/core": "^7.24.4",
     "@emotion/react": "^11.11.4",
     "@emotion/styled": "^11.11.5",
-    "@mui/icons-material": "^5.15.14",
-    "@mui/material": "^5.15.14",
+    "@mui/icons-material": "^5.15.15",
+    "@mui/material": "^5.15.15",
     "@table-library/react-table-library": "4.1.7",
     "@types/imagemin": "^8.0.5",
     "@types/lodash-es": "^4.17.12",
-    "@types/node": "^20.12.2",
+    "@types/node": "^20.12.5",
     "@types/react": "^18.2.74",
-    "@types/react-dom": "^18.2.23",
+    "@types/react-dom": "^18.2.24",
     "@types/react-router-dom": "^5.3.3",
-    "alova": "^2.18.2",
+    "alova": "^2.18.3",
     "async-validator": "^4.2.5",
     "eslint-plugin-prettier": "^5.1.3",
     "history": "^5.3.0",
@@ -50,7 +51,7 @@
     "react-router-dom": "^6.22.3",
     "react-toastify": "^10.0.5",
     "typesafe-i18n": "^5.26.2",
-    "typescript": "^5.4.3"
+    "typescript": "^5.4.4"
   },
   "devDependencies": {
     "@preact/compat": "^17.1.2",
@@ -58,7 +59,7 @@
     "@typescript-eslint/eslint-plugin": "^7.5.0",
     "@typescript-eslint/parser": "^7.5.0",
     "concurrently": "^8.2.2",
-    "eslint": "^8.57.0",
+    "eslint": "^9.0.0",
     "eslint-config-prettier": "^9.1.0",
     "eslint-import-resolver-typescript": "^3.6.1",
     "eslint-plugin-autofix": "^1.1.0",
@@ -68,8 +69,8 @@
     "preact": "^10.20.1",
     "prettier": "^3.2.5",
     "rollup-plugin-visualizer": "^5.12.0",
-    "terser": "^5.30.2",
-    "vite": "^5.2.7",
+    "terser": "^5.30.3",
+    "vite": "^5.2.8",
     "vite-plugin-imagemin": "^0.6.1",
     "vite-tsconfig-paths": "^4.3.2"
   },
