--- mock-api/.gitignore
@@ -1,9 +1,0 @@
-.pnp.*
-.yarn/*
-!.yarn/patches
-!.yarn/plugins
-!.yarn/releases
-!.yarn/sdks
-!.yarn/versions
-uploads/*
-!uploads/README.md
